requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
pdfplumber==0.11.0
google-generativeai==0.7.2
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1